22 a 80
Questão 22 a 80 e algumas correções nas que eu já tinha feito

+++ 4.py
@@ -1,4 +1,4 @@
-nota1 = int(input("Digite a nota 1: "))
-nota2 = int(input("Digite a nota 2: "))
+nota1 = float(input("Digite a nota 1: "))
+nota2 = float(input("Digite a nota 2: "))
 
 print(f"A média do aluno é {(nota1 + nota2) / 2}")
